Added dynamic camera that handles multiple targets
This dynamic camera makes the game look alot more interesting and feel better for multiplayer. I decided against slipt screen since I think this fits better with the game

The camera has funcitons to Add, remove and update from a targets List.
To set the position I just all all targets together and devide by amount of targets.
To set zoom there is a rectangle that uses .Expand( Vector2 ) which expands the rect2 to fit that Vector2. Then .GrowIndividual() to add the margin to the rect aswell so the players are not right on the edge of the viewport.
Then depending on if the rects X or Y is larger ( normalized with screen aspect ration ) it sets the z to the correct size and clamps with min and max zoom bounds.