Reworked player jump to be velocity based
The jump was awful so since the grenades where nice with velocity and stuff I wanted to make the players jump feel good too.
The players jump was very stiff because when the player pressed the jump key it checked if the player was going in a direction and if they where just add that to the jump and the player had zero aircontrol. I booted up terraria just to feel how thier jump felt. Now I have Terraria jump pretty much, velocity based jump.