Instantiate based on game mode
The only thin that exsists in the main scene now is light and a node called "Instantiate Scene" that instantiates everything so I can control what gets instantiated based on the game mode and stage of tournament. What stage of tournament is saved in the gamemanager singleton so the game manager controls what the "Instantiate Scene" node does.