AI controllers
Decided to make seperate files for every enemy, I tried making a more generic EnemyController but it would have been a huge file so This is probably best for the timeframe I have to work with. EnemyController.cs is not a base class for the AI controllers containing all info that all AIcontrollers will use.
* InstantiateController.cs updated to support all enemyControllers. It is also easy to change the tournament setup.